requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
mysqlclient==2.2.7
PyJWT==2.9.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
